Rename witness.ts to witness-proofs.ts, separating witness-proof signing helpers from proof-counting logic
Retains createWitnessProof, signWitnessProofEntry, signWitnessProofEntries,
resolveWitnessParameter, normalizeWitnessThreshold, hasActiveWitnessRequirement,
and validateWitnessParameter in witness-proofs.ts. Removes the now-relocated
countWitnessApprovals/countVerifiedWitnessApprovals and the fetchWitnessProofs
re-export; consumers updated to import from ./witness-proofs.
